+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+- v1.11.0
+	- Added support for FindStuff ploppable rico buildings
+	- Added missing profitability option in mod settings window
 - v1.10.0
 	- Added height and slope to slope mode of the terrain tool
 - v1.9.0